// src/app/models/campaign.model.ts
export interface Campaign {
  ad_id: number;
  xyz_campaign_id: number;
  fb_campaign_id: number;
  age: string;
  gender: string;
  interest: number;
  impressions: number;
  clicks: number;
  spent: number;
  total_conversion: number;
  approved_conversion: number;
  start_date: string;
  end_date: string;
}